+++ crates/maple_cli/src/lib.rs
@@ -4,6 +4,7 @@ mod datastore;
 mod dumb_analyzer;
 mod logger;
 mod paths;
+mod previewer;
 mod process;
 mod recent_files;
 mod stdio_server;